requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-telegram-bot==21.6
google-generativeai==0.7.2
notion-client==2.2.1
trafilatura==1.12.0
pymupdf==1.24.10
httpx==0.27.0
pydantic==2.8.0
pydantic-settings==2.4.0
python-dotenv==1.0.1
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-mock==3.14.0
